Added the completed code
To set up the Data Logger you must ensure that the terminal you are outputting the information to is set to a Baud rate of 38400 and that the Serial port you are using is connected to the corresponding SNAP. After ensuring the connection is set up simply turn on the Data Logger. Once the logger is on you can press any of the four buttons presented. Pressing a button will display the specific button as well as the day, hour, minute, second, and tenth of a second that the button was pressed. Implemented push buttons, an ATtiny2313A, 22pF Ceramic Capacitor, and a 14.7456 MHz Crystal.